+++ lib/canonical_osf1.h
@@ -1,2 +1,2 @@
-  (int)(long)&((struct stringpool2_t *)0)->stringpool_osf1_0,
-  (int)(long)&((struct stringpool2_t *)0)->stringpool_osf1_1,
+  (int)(size_t)&((struct stringpool2_t *)0)->stringpool_osf1_0,
+  (int)(size_t)&((struct stringpool2_t *)0)->stringpool_osf1_1,
